+++ TimelineStatic.cpp
@@ -33,7 +33,7 @@ void CTimelineStatic::DrawItem(LPDRAWITEMSTRUCT lpDrawItemStruct)
 {
 	HDC hdc = CreateCompatibleDC(lpDrawItemStruct->hDC);
 	if (lpDrawItemStruct->itemAction != ODA_DRAWENTIRE) return; //ODS_DEFAULT
-	HBITMAP hbit = CreateCompatibleBitmap(lpDrawItemStruct->hDC, m_motionDetected.size(), 1);
+	HBITMAP hbit = CreateCompatibleBitmap(lpDrawItemStruct->hDC, (int)m_motionDetected.size(), 1);
 	HGDIOBJ hOldBitmap = SelectObject(hdc, hbit);
 	for (int i = 0; i < m_motionDetected.size(); i++) {
 		//red - 255, 0, 0, yellow - 255, 255, 0, white - 255, 255, 255
@@ -43,7 +43,7 @@ void CTimelineStatic::DrawItem(LPDRAWITEMSTRUCT lpDrawItemStruct)
 	StretchBlt(lpDrawItemStruct->hDC, 0, 3,
 		lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left,
 		(lpDrawItemStruct->rcItem.bottom - lpDrawItemStruct->rcItem.top) / 2 - 6,
-		hdc, 0, 0, m_motionDetected.size(), 1, SRCCOPY);
+		hdc, 0, 0, (int)m_motionDetected.size(), 1, SRCCOPY);
 	SelectObject(hdc, hOldBitmap);
 	DeleteObject(hbit);
 	POINT pt;
@@ -53,12 +53,12 @@ void CTimelineStatic::DrawItem(LPDRAWITEMSTRUCT lpDrawItemStruct)
 	r.bottom = (lpDrawItemStruct->rcItem.bottom - lpDrawItemStruct->rcItem.top) / 2;
 	FillRect(lpDrawItemStruct->hDC, &r, (HBRUSH)HOLLOW_BRUSH);
 	if (m_motionDetected.size() != 0) {
-		MoveToEx(lpDrawItemStruct->hDC, (lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) * m_iFrameNum / m_motionDetected.size(), 0, &pt);
-		LineTo(lpDrawItemStruct->hDC, (lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) * m_iFrameNum / m_motionDetected.size(), 3);
-		MoveToEx(lpDrawItemStruct->hDC, (lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) * m_iFrameNum / m_motionDetected.size(), (lpDrawItemStruct->rcItem.bottom - lpDrawItemStruct->rcItem.top) / 2 - 1, &pt);
-		LineTo(lpDrawItemStruct->hDC, (lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) * m_iFrameNum / m_motionDetected.size(), (lpDrawItemStruct->rcItem.bottom - lpDrawItemStruct->rcItem.top) / 2 - 4);
-		RECT rect = { (lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) * m_iFrameNum / m_motionDetected.size(), 3,
-			(lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) * m_iFrameNum / m_motionDetected.size(), (lpDrawItemStruct->rcItem.bottom - lpDrawItemStruct->rcItem.top) / 2 - 3 };
+		MoveToEx(lpDrawItemStruct->hDC, (lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) * m_iFrameNum / (int)m_motionDetected.size(), 0, &pt);
+		LineTo(lpDrawItemStruct->hDC, (lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) * m_iFrameNum / (int)m_motionDetected.size(), 3);
+		MoveToEx(lpDrawItemStruct->hDC, (lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) * m_iFrameNum / (int)m_motionDetected.size(), (lpDrawItemStruct->rcItem.bottom - lpDrawItemStruct->rcItem.top) / 2 - 1, &pt);
+		LineTo(lpDrawItemStruct->hDC, (lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) * m_iFrameNum / (int)m_motionDetected.size(), (lpDrawItemStruct->rcItem.bottom - lpDrawItemStruct->rcItem.top) / 2 - 4);
+		RECT rect = { (lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) * m_iFrameNum / (int)m_motionDetected.size(), 3,
+			(lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) * m_iFrameNum / (int)m_motionDetected.size(), (lpDrawItemStruct->rcItem.bottom - lpDrawItemStruct->rcItem.top) / 2 - 3 };
 		if (rect.left != 0) rect.left--;
 		if (rect.right != (lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left)) rect.right++;
 		InvertRect(lpDrawItemStruct->hDC, &rect);
@@ -69,20 +69,20 @@ void CTimelineStatic::DrawItem(LPDRAWITEMSTRUCT lpDrawItemStruct)
 	//HGDIOBJ hOldObj = SelectObject(hdc, (HBRUSH)WHITE_BRUSH); //WHITE_BRUSH and BLACK_PEN are defaults
 	r = lpDrawItemStruct->rcItem;
 	r.top += 5;
-	ExtTextOut(lpDrawItemStruct->hDC, (lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) / 2, 0, 0, &r, buf, _tcslen(buf), NULL);
+	ExtTextOut(lpDrawItemStruct->hDC, (lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left) / 2, 0, 0, &r, buf, (UINT)_tcslen(buf), NULL);
 	//SelectObject(hdc, hOldObj);
 
-	hbit = CreateCompatibleBitmap(lpDrawItemStruct->hDC, m_oscillation.size(), 30);
+	hbit = CreateCompatibleBitmap(lpDrawItemStruct->hDC, (int)m_oscillation.size(), 30);
 	hOldBitmap = SelectObject(hdc, hbit);
-	r.left = 0; r.top = 0; r.right = m_oscillation.size(); r.bottom = 30;
+	r.left = 0; r.top = 0; r.right = (LONG)m_oscillation.size(); r.bottom = 30;
 	FillRect(hdc, &r, (HBRUSH)BLACK_BRUSH);
 	for (int i = 0; i < m_oscillation.size(); i++) {
-		SetPixel(hdc, i, 30 * m_oscillation[i] / m_dMaxOscillation, RGB(255, 255, 255));
+		SetPixel(hdc, i, (int)(30 * m_oscillation[i] / m_dMaxOscillation), RGB(255, 255, 255));
 	}
 	StretchBlt(lpDrawItemStruct->hDC, 0, (lpDrawItemStruct->rcItem.bottom - lpDrawItemStruct->rcItem.top) / 2,
 		lpDrawItemStruct->rcItem.right - lpDrawItemStruct->rcItem.left,
 		lpDrawItemStruct->rcItem.bottom - (lpDrawItemStruct->rcItem.bottom - lpDrawItemStruct->rcItem.top) / 2,
-		hdc, 0, 0, m_oscillation.size(), 30, SRCCOPY);
+		hdc, 0, 0, (int)m_oscillation.size(), 30, SRCCOPY);
 	SelectObject(hdc, hOldBitmap);
 	DeleteObject(hbit);
 	DeleteDC(hdc);
@@ -94,7 +94,7 @@ void CTimelineStatic::OnLButtonDown(UINT nFlags, CPoint point)
 	RECT rect;
 	GetWindowRect(&rect);
 	ScreenToClient(&rect);
-	GetParent()->PostMessage(WM_CUSTOM_MOVEVIDEO, 0, m_iFrameNum = m_motionDetected.size() * point.x / (rect.right - rect.left + 1));
+	GetParent()->PostMessage(WM_CUSTOM_MOVEVIDEO, 0, m_iFrameNum = (int)m_motionDetected.size() * point.x / (rect.right - rect.left + 1));
 	Invalidate();
 	CStatic::OnLButtonDown(nFlags, point);
 }
